+++ packages/client/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this npm module are documented in this file.
 
+## [9.2.1] - 2025-01-31
+
+- Moved workerFactory from `monaco-editor-wrapper` to `monaco-languageclient`.
+
 ## [9.2.0] - 2025-01-31
 
 - Update to monaco-vscode-api v13 [#836](https://github.com/TypeFox/monaco-languageclient/pull/829)